requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
aiohttp==3.9.1
sqlite3
asyncio
datetime
typing
pydantic==2.5.0